sign transaction reduce and cancel, fix total value display

+++ Data/Oracle/CriptoDTO.cs
@@ -4,7 +4,7 @@ namespace Data.Oracle
 {
     public class CriptoDTO
     {
-        public decimal TotalBid { get; set; }
+        public ulong TotalBid { get; set; }
         public long Time { get; set; }
         public DateTime DTTime { get; set; }
     }
